1) peakdet.py has been renamed as peak_detection.py 2) Only functions which are going to be used 3) Tests of unused functions have been deleted 4) Both findpeaks1D.py and peakdetection.py have been deleted
