+++ numina/array/peaks/tests/test_peakdet.py
@@ -1,18 +1,9 @@
 import sys
-
-import numpy
 import pytest
-
-from ..findpeaks1D import findPeaks_spectrum, refinePeaks_spectrum
-
-from ..peakdet import generate_kernel
-from ..peakdet import find_peaks_index1
-from ..peakdet import find_peaks_index2
-from ..peakdet import refine_peaks1, refine_peaks2, refine_peaks3
 from .._kernels import kernel_peak_function
 
-@pytest.mark.skipif(sys.version_info < (3,0),
-                    reason="requires python3")
+
+@pytest.mark.skipif(sys.version_info < (3,0),reason="requires python3")
 def test_pycapsule():
     m = kernel_peak_function(20)
     # This seems to be the only way to check that something is a PyCapsule
@@ -20,16 +11,18 @@ def test_pycapsule():
     assert sm == 'PyCapsule'
 
 
-@pytest.mark.skipif(sys.version_info >= (3,0),
-                    reason="requires python2")
+@pytest.mark.skipif(sys.version_info >= (3,0),reason="requires python2")
 def test_pycobject():
     m = kernel_peak_function(20)
     # This seems to be the only way to check that something is a PyCapsule
     sm = type(m).__name__
     assert sm == 'PyCObject'
 
+
 @pytest.mark.parametrize("window", [3, 5, 7, 9])
 def test_generate_kernel(benchmark, window):
+    from numina.array.peaks.peak_detection import generate_kernel
+
     result = benchmark(generate_kernel, window)
 
     assert result.shape == (3, window)
@@ -37,6 +30,7 @@ def test_generate_kernel(benchmark, window):
 
 @pytest.fixture(scope="module")
 def spectrum():
+    import numpy
     yl = [
         1.929785E-16,
         1.947651E-16,
@@ -142,52 +136,20 @@ def spectrum():
     return numpy.array(yl)
 
 
-def test_peak_finding_base(benchmark, spectrum):
-    peakin = [8, 56, 88]
-    result = benchmark(findPeaks_spectrum, spectrum, 5, 0.5e-15)
-    assert numpy.allclose(peakin, result)
-
-
-def test_peak_finding_v1(benchmark, spectrum):
-    # Reference
-    peakin = [8, 56, 88]
-    result = benchmark(find_peaks_index1, spectrum, threshold=0.5e-15, window=5)
-    assert numpy.allclose(peakin, result)
-
-
 def test_peak_finding_v2(benchmark, spectrum):
-    # Reference
+    from numina.array.peaks.peak_detection import find_peaks_index
+    import numpy
+
     peakin = [8, 56, 88]
-    result = benchmark(find_peaks_index2, spectrum, threshold=0.5e-15, window=5)
+    result = benchmark(find_peaks_index, spectrum, threshold=0.5e-15, window=5)
     assert numpy.allclose(peakin, result)
 
-
-def test_peak_refine_original(benchmark, spectrum):
-    peakin = findPeaks_spectrum(spectrum, 5, 0.5e-15)
-    peakpos = benchmark(refinePeaks_spectrum, spectrum, peakin, 5, method=1)
-
-    assert peakpos is not None
-
-
-def test_peak_refine_loop(benchmark, spectrum):
-
-    peakin = findPeaks_spectrum(spectrum, 5, 0.5e-15)
-    peakpos = benchmark(refine_peaks1, spectrum, peakin, 5)
-
-    assert peakpos is not None
-
-
-def test_peak_refine_no_loop(benchmark, spectrum):
-
-    peakin = findPeaks_spectrum(spectrum, 5, 0.5e-15)
-    peakpos = benchmark(refine_peaks2, spectrum, peakin, 5)
-
-    assert peakpos is not None
 
 
 def test_peak_refine_no_loop_compW(benchmark, spectrum):
+    from numina.array.peaks.peak_detection import find_peaks_index, refine_peaks
 
-    peakin = findPeaks_spectrum(spectrum, 5, 0.5e-15)
-    peakpos = benchmark(refine_peaks3, spectrum, peakin, 5)
+    peakin = find_peaks_index(spectrum, 5, 0.5e-15)
+    peakpos = benchmark(refine_peaks, spectrum, peakin, 5)
 
     assert peakpos is not None
